GW Setup
The main commit in preparation to start the GW, this commit includes "basic" views showing the pixel buffer in a different format separated from the emulator (by using its initial buffer). This allows a simple representation of the pixels that allows for comparisons. This setup includes a few views such as the gameboy view and a view showing my own "pixel buffer" + also using it to check out individual pixels (with surroundings)

[MOD] SDL - edited version
[MOD] ImGui - edited version
[ADD] EView - EViewHolder 
[ADD] Renderer
[ADD] TetrisSystem
[ADD] Views
[ADD] Smaller util files
[DEL] Demo file