Tetris System - Calculate Best Move
Following previous research, I've added the score calculation.
I've created an additional view that can showcase the best move according to the algorithm and using a variable depth.a
The move might not seem to be the best according to human eyes, however. The algorithm works in strange, yet can get a very good end result (according to the source).