Tetris System - Extended Block Prediction
I've rewritten the Block Prediction view to allow for multiple blocks. This would allow me to see if there is an issue with the algorithm or with my own code when it comes to the wrong best move prediction. It seems the cause is in the best move prediction itself as I can manually find better moves with a better score that would make way more sense.