Tetris System - Best Move Prediction Fix
A very dumb mistake and also a dumb move linking was the cause.
The dumb mistake was that the moves it did beyond the first piece, was only the ones of the FINAL move of the previous piece. Bless "insert".
Another dumb issue then found by allowing piece colors, was that for the movesDepth, there were some issues, instead of doing it like that, I added a prev move pointer which made things easier.