Tetris System - Fixed Input
Further improvements to the system including wider score zone checking, improved input handling, fps limitter, etc.

I've put an FPS limiter on the system to limit it to the same speed as the Gameboy. While putting the fps of the system higher, could improve score and level (which could also be sheer luck), it is not worth it as performance of the gameboy somehow looks worse and sometimes buttons don't get fully pressed.
The option stays there for testing in case someone else picks it up.

The system is now in full control, always and keeps replaying level 1 (starts from 0).