Tetris System - Moved early return
Moved an early return in function. There could still be a move for other previous piece moves. Rather check if after checking all those moves if it is empty. 
An early exit there is almost necessary to make sure it doesn't "crash" at the end as it is expected there is a move at the final piece.